+++ dapp/package.json
@@ -38,6 +38,8 @@
 		"@radix-ui/react-separator": "^1.1.0",
 		"@radix-ui/react-slot": "^1.0.2",
 		"@radix-ui/react-tooltip": "^1.1.3",
+		"@reown/appkit": "^1.1.6",
+		"@reown/appkit-adapter-wagmi": "^1.1.6",
 		"@t3-oss/env-nextjs": "^0.9.2",
 		"@tanstack/react-query": "4.x",
 		"@trpc/client": "^10.45.2",
@@ -80,7 +82,6 @@
 		"@typescript-eslint/eslint-plugin": "^7.0.2",
 		"@typescript-eslint/parser": "^7.0.2",
 		"@wagmi/core": "^2.13.4",
-		"@web3modal/wagmi": "^4.1.1",
 		"babel-loader": "^9.1.3",
 		"babel-plugin-macros": "^3.1.0",
 		"babel-plugin-twin": "^1.1.0",
